- refactoring architype view - add idle

+++ wiki/src/main/java/com/dnd5e/wiki/model/hero/ArchetypeTrait.java
@@ -25,7 +25,7 @@ public class ArchetypeTrait {
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Integer id;
 	private String name;
-	private int level;
+	private byte level;
 	@Column(columnDefinition = "TEXT")
 	private String description;
